store/tikv:remove codec from snapshot.go
