+++ store/tikv/tests/async_commit_fail_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/pingcap/parser/terror"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/store/tikv"
+	tikverr "github.com/pingcap/tidb/store/tikv/kv"
 	"github.com/pingcap/tidb/store/tikv/util"
 )
 
@@ -102,7 +103,9 @@ func (s *testAsyncCommitFailSuite) TestAsyncCommitPrewriteCancelled(c *C) {
 	ctx := context.WithValue(context.Background(), util.SessionID, uint64(1))
 	err = t1.Commit(ctx)
 	c.Assert(err, NotNil)
-	c.Assert(kv.ErrWriteConflict.Equal(err), IsTrue, Commentf("%s", errors.ErrorStack(err)))
+	e, ok := errors.Cause(err).(*tikverr.UnExtractKeyErr)
+	c.Assert(ok, IsTrue, Commentf("%s", errors.ErrorStack(err)))
+	c.Assert(e.GetConflict(), NotNil, Commentf("%s", errors.ErrorStack(err)))
 }
 
 func (s *testAsyncCommitFailSuite) TestPointGetWithAsyncCommit(c *C) {
